Nametag

+++ src/java/net/minecraft/block/BlockAir.java
@@ -18,7 +18,7 @@ public class BlockAir extends Block {
 
     public static void setLightOpacity(Block p_setLightOpacity_0_, int p_setLightOpacity_1_) {
         if (!mapOriginalOpacity.containsKey(p_setLightOpacity_0_)) {
-            mapOriginalOpacity.put(p_setLightOpacity_0_, Integer.valueOf(p_setLightOpacity_0_.lightOpacity));
+            mapOriginalOpacity.put(p_setLightOpacity_0_, p_setLightOpacity_0_.lightOpacity);
         }
 
         p_setLightOpacity_0_.lightOpacity = p_setLightOpacity_1_;
@@ -26,7 +26,7 @@ public class BlockAir extends Block {
 
     public static void restoreLightOpacity(Block p_restoreLightOpacity_0_) {
         if (mapOriginalOpacity.containsKey(p_restoreLightOpacity_0_)) {
-            int i = ((Integer) mapOriginalOpacity.get(p_restoreLightOpacity_0_)).intValue();
+            int i = (Integer) mapOriginalOpacity.get(p_restoreLightOpacity_0_);
             setLightOpacity(p_restoreLightOpacity_0_, i);
         }
     }
