+++ src/java/net/minecraft/block/BlockBanner.java
@@ -191,7 +191,7 @@ public class BlockBanner extends BlockContainer {
 
     public static class BlockBannerStanding extends BlockBanner {
         public BlockBannerStanding() {
-            this.setDefaultState(this.blockState.getBaseState().withProperty(ROTATION, Integer.valueOf(0)));
+            this.setDefaultState(this.blockState.getBaseState().withProperty(ROTATION, 0));
         }
 
         public void onNeighborBlockChange(World worldIn, BlockPos pos, IBlockState state, Block neighborBlock) {
@@ -204,11 +204,11 @@ public class BlockBanner extends BlockContainer {
         }
 
         public IBlockState getStateFromMeta(int meta) {
-            return this.getDefaultState().withProperty(ROTATION, Integer.valueOf(meta));
+            return this.getDefaultState().withProperty(ROTATION, meta);
         }
 
         public int getMetaFromState(IBlockState state) {
-            return state.getValue(ROTATION).intValue();
+            return state.getValue(ROTATION);
         }
 
         protected BlockState createBlockState() {
