+++ src/java/net/minecraft/block/BlockCactus.java
@@ -18,7 +18,7 @@ public class BlockCactus extends Block {
 
     protected BlockCactus() {
         super(Material.cactus);
-        this.setDefaultState(this.blockState.getBaseState().withProperty(AGE, Integer.valueOf(0)));
+        this.setDefaultState(this.blockState.getBaseState().withProperty(AGE, 0));
         this.setTickRandomly(true);
         this.setCreativeTab(CreativeTabs.tabDecorations);
     }
@@ -33,15 +33,15 @@ public class BlockCactus extends Block {
             }
 
             if (i < 3) {
-                int j = state.getValue(AGE).intValue();
+                int j = state.getValue(AGE);
 
                 if (j == 15) {
                     worldIn.setBlockState(blockpos, this.getDefaultState());
-                    IBlockState iblockstate = state.withProperty(AGE, Integer.valueOf(0));
+                    IBlockState iblockstate = state.withProperty(AGE, 0);
                     worldIn.setBlockState(pos, iblockstate, 4);
                     this.onNeighborBlockChange(worldIn, blockpos, iblockstate, this);
                 } else {
-                    worldIn.setBlockState(pos, state.withProperty(AGE, Integer.valueOf(j + 1)), 4);
+                    worldIn.setBlockState(pos, state.withProperty(AGE, j + 1), 4);
                 }
             }
         }
@@ -107,14 +107,14 @@ public class BlockCactus extends Block {
      * Convert the given metadata into a BlockState for this Block
      */
     public IBlockState getStateFromMeta(int meta) {
-        return this.getDefaultState().withProperty(AGE, Integer.valueOf(meta));
+        return this.getDefaultState().withProperty(AGE, meta);
     }
 
     /**
      * Convert the BlockState into the correct metadata value
      */
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(AGE).intValue();
+        return state.getValue(AGE);
     }
 
     protected BlockState createBlockState() {
