+++ src/java/net/minecraft/block/BlockCrops.java
@@ -19,7 +19,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
     public static final PropertyInteger AGE = PropertyInteger.create("age", 0, 7);
 
     protected BlockCrops() {
-        this.setDefaultState(this.blockState.getBaseState().withProperty(AGE, Integer.valueOf(0)));
+        this.setDefaultState(this.blockState.getBaseState().withProperty(AGE, 0));
         this.setTickRandomly(true);
         float f = 0.5F;
         this.setBlockBounds(0.5F - f, 0.0F, 0.5F - f, 0.5F + f, 0.25F, 0.5F + f);
@@ -41,7 +41,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
                 if (iblockstate.getBlock() == Blocks.farmland) {
                     f1 = 1.0F;
 
-                    if (iblockstate.getValue(BlockFarmland.MOISTURE).intValue() > 0) {
+                    if (iblockstate.getValue(BlockFarmland.MOISTURE) > 0) {
                         f1 = 3.0F;
                     }
                 }
@@ -85,26 +85,26 @@ public class BlockCrops extends BlockBush implements IGrowable {
         super.updateTick(worldIn, pos, state, rand);
 
         if (worldIn.getLightFromNeighbors(pos.up()) >= 9) {
-            int i = state.getValue(AGE).intValue();
+            int i = state.getValue(AGE);
 
             if (i < 7) {
                 float f = getGrowthChance(this, worldIn, pos);
 
                 if (rand.nextInt((int) (25.0F / f) + 1) == 0) {
-                    worldIn.setBlockState(pos, state.withProperty(AGE, Integer.valueOf(i + 1)), 2);
+                    worldIn.setBlockState(pos, state.withProperty(AGE, i + 1), 2);
                 }
             }
         }
     }
 
     public void grow(World worldIn, BlockPos pos, IBlockState state) {
-        int i = state.getValue(AGE).intValue() + MathHelper.getRandomIntegerInRange(worldIn.rand, 2, 5);
+        int i = state.getValue(AGE) + MathHelper.getRandomIntegerInRange(worldIn.rand, 2, 5);
 
         if (i > 7) {
             i = 7;
         }
 
-        worldIn.setBlockState(pos, state.withProperty(AGE, Integer.valueOf(i)), 2);
+        worldIn.setBlockState(pos, state.withProperty(AGE, i), 2);
     }
 
     public boolean canBlockStay(World worldIn, BlockPos pos, IBlockState state) {
@@ -126,7 +126,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
         super.dropBlockAsItemWithChance(worldIn, pos, state, chance, 0);
 
         if (!worldIn.isRemote) {
-            int i = state.getValue(AGE).intValue();
+            int i = state.getValue(AGE);
 
             if (i >= 7) {
                 int j = 3 + fortune;
@@ -144,7 +144,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
      * Get the Item that this Block should drop when harvested.
      */
     public Item getItemDropped(IBlockState state, Random rand, int fortune) {
-        return state.getValue(AGE).intValue() == 7 ? this.getCrop() : this.getSeed();
+        return state.getValue(AGE) == 7 ? this.getCrop() : this.getSeed();
     }
 
     public Item getItem(World worldIn, BlockPos pos) {
@@ -155,7 +155,7 @@ public class BlockCrops extends BlockBush implements IGrowable {
      * Whether this IGrowable can grow
      */
     public boolean canGrow(World worldIn, BlockPos pos, IBlockState state, boolean isClient) {
-        return state.getValue(AGE).intValue() < 7;
+        return state.getValue(AGE) < 7;
     }
 
     public boolean canUseBonemeal(World worldIn, Random rand, BlockPos pos, IBlockState state) {
@@ -170,14 +170,14 @@ public class BlockCrops extends BlockBush implements IGrowable {
      * Convert the given metadata into a BlockState for this Block
      */
     public IBlockState getStateFromMeta(int meta) {
-        return this.getDefaultState().withProperty(AGE, Integer.valueOf(meta));
+        return this.getDefaultState().withProperty(AGE, meta);
     }
 
     /**
      * Convert the BlockState into the correct metadata value
      */
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(AGE).intValue();
+        return state.getValue(AGE);
     }
 
     protected BlockState createBlockState() {
