+++ src/java/net/minecraft/block/BlockLiquid.java
@@ -20,7 +20,7 @@ public abstract class BlockLiquid extends Block {
 
     protected BlockLiquid(Material materialIn) {
         super(materialIn);
-        this.setDefaultState(this.blockState.getBaseState().withProperty(LEVEL, Integer.valueOf(0)));
+        this.setDefaultState(this.blockState.getBaseState().withProperty(LEVEL, 0));
         this.setBlockBounds(0.0F, 0.0F, 0.0F, 1.0F, 1.0F, 1.0F);
         this.setTickRandomly(true);
     }
@@ -70,7 +70,7 @@ public abstract class BlockLiquid extends Block {
     }
 
     protected int getLevel(IBlockAccess worldIn, BlockPos pos) {
-        return worldIn.getBlockState(pos).getBlock().getMaterial() == this.blockMaterial ? worldIn.getBlockState(pos).getValue(LEVEL).intValue() : -1;
+        return worldIn.getBlockState(pos).getBlock().getMaterial() == this.blockMaterial ? worldIn.getBlockState(pos).getValue(LEVEL) : -1;
     }
 
     protected int getEffectiveFlowDecay(IBlockAccess worldIn, BlockPos pos) {
@@ -90,7 +90,7 @@ public abstract class BlockLiquid extends Block {
     }
 
     public boolean canCollideCheck(IBlockState state, boolean hitIfLiquid) {
-        return hitIfLiquid && state.getValue(LEVEL).intValue() == 0;
+        return hitIfLiquid && state.getValue(LEVEL) == 0;
     }
 
     /**
@@ -169,7 +169,7 @@ public abstract class BlockLiquid extends Block {
             }
         }
 
-        if (worldIn.getBlockState(pos).getValue(LEVEL).intValue() >= 8) {
+        if (worldIn.getBlockState(pos).getValue(LEVEL) >= 8) {
             for (EnumFacing enumfacing1 : EnumFacing.Plane.HORIZONTAL) {
                 BlockPos blockpos1 = pos.offset(enumfacing1);
 
@@ -214,7 +214,7 @@ public abstract class BlockLiquid extends Block {
         double d2 = pos.getZ();
 
         if (this.blockMaterial == Material.water) {
-            int i = state.getValue(LEVEL).intValue();
+            int i = state.getValue(LEVEL);
 
             if (i > 0 && i < 8) {
                 if (rand.nextInt(64) == 0) {
@@ -281,13 +281,13 @@ public abstract class BlockLiquid extends Block {
             if (flag) {
                 Integer integer = state.getValue(LEVEL);
 
-                if (integer.intValue() == 0) {
+                if (integer == 0) {
                     worldIn.setBlockState(pos, Blocks.obsidian.getDefaultState());
                     this.triggerMixEffects(worldIn, pos);
                     return true;
                 }
 
-                if (integer.intValue() <= 4) {
+                if (integer <= 4) {
                     worldIn.setBlockState(pos, Blocks.cobblestone.getDefaultState());
                     this.triggerMixEffects(worldIn, pos);
                     return true;
@@ -313,14 +313,14 @@ public abstract class BlockLiquid extends Block {
      * Convert the given metadata into a BlockState for this Block
      */
     public IBlockState getStateFromMeta(int meta) {
-        return this.getDefaultState().withProperty(LEVEL, Integer.valueOf(meta));
+        return this.getDefaultState().withProperty(LEVEL, meta);
     }
 
     /**
      * Convert the BlockState into the correct metadata value
      */
     public int getMetaFromState(IBlockState state) {
-        return state.getValue(LEVEL).intValue();
+        return state.getValue(LEVEL);
     }
 
     protected BlockState createBlockState() {
