+++ src/java/net/minecraft/block/BlockPotato.java
@@ -23,7 +23,7 @@ public class BlockPotato extends BlockCrops {
         super.dropBlockAsItemWithChance(worldIn, pos, state, chance, fortune);
 
         if (!worldIn.isRemote) {
-            if (state.getValue(AGE).intValue() >= 7 && worldIn.rand.nextInt(50) == 0) {
+            if (state.getValue(AGE) >= 7 && worldIn.rand.nextInt(50) == 0) {
                 spawnAsEntity(worldIn, pos, new ItemStack(Items.poisonous_potato));
             }
         }
